LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-column-axis-self-baseline-synthesized-003-expected.txt

Horizontal grid and item with relative height
É É ÉÉ
VerticalLR grid and item with relative width
É É ÉÉ
VerticalRL grid and item with rlative width
É É ÉÉ

FAIL .grid 1 assert_equals: 
<div class="grid alignItemsBaseline verticalRL">
  <div class="firstRowFirstColumn relativeWidth" data-offset-x="200" data-offset-y="0" data-expected-width="100" data-expected-height="100"></div>
  <div class="firstRowSecondColumn bigFont paddingRight" data-offset-x="140" data-offset-y="100" data-expected-width="120" data-expected-height="200">É É ÉÉ</div>
  <div class="autoRowAutoColumnSpanning2 width25"></div>
</div>
offsetLeft expected 140 but got 180
FAIL .grid 2 assert_equals: 
<div class="grid alignItemsBaseline verticalLR">
  <div class="firstRowFirstColumn relativeWidth" data-offset-x="30" data-offset-y="0" data-expected-width="100" data-expected-height="100"></div>
  <div class="firstRowSecondColumn bigFont paddingLeft" data-offset-x="0" data-offset-y="100" data-expected-width="120" data-expected-height="200">É É ÉÉ</div>
  <div class="autoRowAutoColumnSpanning2 width25"></div>
</div>
offsetLeft expected 30 but got 0
FAIL .grid 3 assert_equals: 
<div class="grid width300 alignItemsBaseline">
  <div class="firstRowFirstColumn relativeHeight" data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="100"></div>
  <div class="firstRowSecondColumn bigFont" data-offset-x="100" data-offset-y="60" data-expected-width="200" data-expected-height="100">É É ÉÉ</div>
  <div class="autoRowAutoColumnSpanning2 height25"></div>
</div>
offsetTop expected 60 but got 0
